Fixed All pairs Algorithm and Improve performance

+++ include/data_structures/nbody.h
@@ -1,13 +1,9 @@
 #ifndef NBODY_H
 #define NBODY_H
 
-#include "body.h"
-#include <vector>
-
 class NBody {
  public:
-  virtual void Update(double dt) = 0;  // Pure virtual method
-
+  virtual void Update(float dt) = 0;  // Pure virtual method
 };
 
 #endif
